Added test support and added rbac_domain functions.
